+++ src/SafeOrbit/Memory/SafeString/ISafeString.cs
@@ -1,31 +1,22 @@
 ﻿using System;
 using System.Threading.Tasks;
-using SafeOrbit.Common;
 using SafeOrbit.Memory.SafeStringServices.Text;
 
 namespace SafeOrbit.Memory
 {
-    public interface ISafeString : IAsyncEquatable<string>, IAsyncEquatable<ISafeString>,
+    public interface ISafeString : IReadOnlySafeString,
         IAsyncDeepCloneable<ISafeString>, IShallowCloneable<ISafeString>,
         IDisposable
     {
-        bool IsEmpty { get; }
-        bool IsDisposed { get; }
-
         /// <summary>
-        ///     Gets the number of characters in the current SafeString
-        ///     Disposed SafeString returns 0
-        /// </summary>
-        int Length { get; }
-
-        Task<char> RevealDecryptedCharAsync(int index);
-        Task<byte[]> RevealDecryptedBytesAsync();
-        ISafeBytes GetAsSafeBytes(int index);
-
-        /// <summary>
-        ///     Appends a <see cref="String" /> that's already revealed in the memory.
+        ///     Appends a <see cref="char" /> that's already revealed in the memory.
         /// </summary>
         Task AppendAsync(char ch);
+
+        /// <summary>
+        ///     Appends a <see cref="string" /> that's already revealed in the memory.
+        /// </summary>
+        Task AppendAsync(string text);
 
         /// <summary>
         ///     Append bytes of a single character
@@ -34,20 +25,18 @@ namespace SafeOrbit.Memory
         /// <param name="encoding">Encoding type that bytes are stored. Default = little endian UTF16</param>
         Task AppendAsync(ISafeBytes character, Encoding encoding = Encoding.Utf16LittleEndian);
 
+        /// <summary>
+        ///     Appends encrypted <paramref name="safeString"/> at the end of the current <see cref="ISafeString"/> instance.
+        /// </summary>
         Task AppendAsync(ISafeString safeString);
 
         /// <summary>
-        ///     Adds string that's not important for memory leaks. Will be seen in memory.
-        /// </summary>
-        /// <param name="text">String to append</param>
-        Task AppendAsync(string text);
-
-        /// <summary>
-        ///     Appends the default line terminator (Environment.NewLine) to the end of the
-        ///     current SafeString object
+        ///     Appends the default line terminator to the end of the current <see cref="ISafeString"/> instance.
         /// </summary>
         Task AppendLineAsync();
+
         Task InsertAsync(int index, char character);
+
         Task InsertAsync(int position, ISafeBytes character, Encoding encoding = Encoding.Utf16LittleEndian);
 
         /// <param name="startIndex">Position of char to remove</param>
@@ -58,7 +47,5 @@ namespace SafeOrbit.Memory
         ///     Removes all data
         /// </summary>
         void Clear();
-
-        int GetHashCode();
     }
 }