+++ src/SafeOrbit/CHANGELOG.md
@@ -32,6 +32,7 @@
   - `AppendMany()` to append multiple bytes to `SafeBytes`
   - Added constructor that writes the given binary to `SafeMemoryStream`
   - Added `RevealDecryptedBytesAsync` in `SafeString`
+  - Added immutable interfaces `IReadOnlySafeString` and `IReadOnlySafeBytes`
 - **Changed**
   - ❗ Replaced sync methods with async variants.
   - Simplified builds: targets only `netstandard2.0`, `netstandard1.6` and `NET4.5`.